test: extracts torrent and logos tests to separate files
